+++ resource/template/fileStat_posix.go
@@ -1,0 +1,32 @@
+// +build !windows
+
+package template
+
+import (
+	"crypto/md5"
+	"errors"
+	"fmt"
+	"io"
+	"os"
+	"syscall"
+)
+
+// fileStat return a fileInfo describing the named file.
+func fileStat(name string) (fi fileInfo, err error) {
+	if isFileExist(name) {
+		f, err := os.Open(name)
+		if err != nil {
+			return fi, err
+		}
+		defer f.Close()
+		stats, _ := f.Stat()
+		fi.Uid = stats.Sys().(*syscall.Stat_t).Uid
+		fi.Gid = stats.Sys().(*syscall.Stat_t).Gid
+		fi.Mode = stats.Mode()
+		h := md5.New()
+		io.Copy(h, f)
+		fi.Md5 = fmt.Sprintf("%x", h.Sum(nil))
+		return fi, nil
+	}
+	return fi, errors.New("File not found")
+}
