+++ resource/template/util.go
@@ -1,14 +1,10 @@
 package template
 
 import (
-	"crypto/md5"
-	"errors"
 	"fmt"
-	"io"
 	"os"
 	"path"
 	"path/filepath"
-	"syscall"
 
 	"github.com/kelseyhightower/confd/log"
 )
@@ -35,27 +31,6 @@ func isFileExist(fpath string) bool {
 		return false
 	}
 	return true
-}
-
-// fileStat return a fileInfo describing the named file.
-func fileStat(name string) (fi fileInfo, err error) {
-	if isFileExist(name) {
-		f, err := os.Open(name)
-		if err != nil {
-			return fi, err
-		}
-		defer f.Close()
-		stats, _ := f.Stat()
-		fi.Uid = stats.Sys().(*syscall.Stat_t).Uid
-		fi.Gid = stats.Sys().(*syscall.Stat_t).Gid
-		fi.Mode = stats.Mode()
-		h := md5.New()
-		io.Copy(h, f)
-		fi.Md5 = fmt.Sprintf("%x", h.Sum(nil))
-		return fi, nil
-	} else {
-		return fi, errors.New("File not found")
-	}
 }
 
 // sameConfig reports whether src and dest config files are equal.
